libkernel: CpuOps: make interrupt flags a associated type
Rather than hard-coding the interrupt flags as a `usize`. Allow each CPU
architecture to define it's own interrupt flags type via an associated
type.

+++ src/arch/arm64/memory/heap.rs
@@ -30,7 +30,7 @@ impl SlabGetter<ArchImpl, PgAllocGetter, PageOffsetTranslator> for StaticSlabGet
 }
 
 pub struct PerCpuCache {
-    flags: usize,
+    flags: u64,
 }
 
 impl PerCpuCache {
